refactor: 修复审计问题，提取公共逻辑 & 清理代码规范
Summary
  - 提取星级过滤器逻辑为 useRarityFilters composable，消除 3 处重复代码
  - 提取 bestChoices/allChoices 共享排序逻辑为 _sortedChoices computed
  - 重命名单字母变量（s → stat, r → rec）符合命名规范
  - 为 _recomputeChoices 添加算法说明注释
  - 删除未使用的 getFarmingRecommendation 死代码

Changes
  新增
  - `useRarityFilters.ts` — 星级过滤器 composable，从 profile 加载/保存过滤设置

  修改
  - `WeaponOverview.vue` — 使用 useRarityFilters 替代内联 watcher，移除 watch import
  - `matrix-planner.vue` — 使用 useRarityFilters，重命名 s → stat
  - `treasure-matrix.vue` — 使用 useRarityFilters，重命名 r → rec
  - `useMatrixPlanner.ts` — 提取 _sortedChoices，重命名 s → stat，添加算法注释

  删除
  - `useProfiles.ts` — 移除未使用的 getFarmingRecommendation 函数及其导出
