+++ frontend/src/components/WeaponOverview.vue
@@ -99,20 +99,20 @@
 </template>
 
 <script lang="ts" setup>
-import { computed, ref, watch } from 'vue'
+import { computed, ref } from 'vue'
 import ItemIcon from '@/components/ItemIcon.vue'
 import { useProfiles } from '@/composables/useProfiles'
+import { useRarityFilters } from '@/composables/useRarityFilters'
 import { useStaticData } from '@/utils/gameData/staticData'
 import { getGemTagName } from '@/utils/gameData/weapon'
 
 const { weaponTypes, weaponsMap } = useStaticData()
 const {
-  activeProfile,
   treasureMatrix,
   addTreasureMatrixEntry,
   removeTreasureMatrixEntry,
-  updateWeaponOverviewFilters,
 } = useProfiles()
+const { selectedRarities } = useRarityFilters()
 
 // 当前激活的武器ID（用于显示弹出提示）
 const activeWeaponId = ref<string | null>(null)
@@ -126,53 +126,6 @@ const totalCount = computed(() =>
 )
 
 const ownedCount = computed(() => ownedWeaponIds.value.length)
-
-// 星级过滤器
-const selectedRarities = ref<string[]>(['3', '4', '5', '6'])
-
-// 从profile加载过滤器设置
-watch(
-  () => activeProfile.value.weapon_overview_filters,
-  (filters) => {
-    if (filters) {
-      const newRarities: string[] = []
-      if (filters['3star']) newRarities.push('3')
-      if (filters['4star']) newRarities.push('4')
-      if (filters['5star']) newRarities.push('5')
-      if (filters['6star']) newRarities.push('6')
-
-      // 只有在实际不同时才更新，避免循环
-      const current = selectedRarities.value.toSorted().join(',')
-      const updated = newRarities.toSorted().join(',')
-      if (current !== updated) {
-        selectedRarities.value = newRarities
-      }
-    }
-  },
-  { immediate: true },
-)
-
-// 保存过滤器设置
-watch(
-  selectedRarities,
-  async (newValue, oldValue) => {
-    // 避免初始化时触发
-    if (!oldValue) return
-
-    // 检查是否真的有变化
-    const oldSorted = oldValue.toSorted().join(',')
-    const newSorted = newValue.toSorted().join(',')
-    if (oldSorted === newSorted) return
-
-    await updateWeaponOverviewFilters({
-      '3star': newValue.includes('3'),
-      '4star': newValue.includes('4'),
-      '5star': newValue.includes('5'),
-      '6star': newValue.includes('6'),
-    })
-  },
-  { deep: true },
-)
 
 // 过滤后的武器类型列表
 const filteredWeaponTypes = computed(() => {
